all index members for GetIndex/SetIndex MUST be either #NO_COPY or protected if you duplicate an item and its index gets copied, the new guy gets the wrong index was happening with textures in virtenv.

+++ src/temt/ta/ta_datatable.h
@@ -133,7 +133,7 @@ public:
 
   String		desc; // #NO_SAVE_EMPTY #EDIT_DIALOG optional description to help in documenting the use of this column
   ColFlags		col_flags; // #APPLY_IMMED flags for this column to indicate specific properties 
-  short			col_idx; // #READ_ONLY #SHOW #NO_SAVE the index of this column in the table
+  int			col_idx; // #READ_ONLY #SHOW #NO_SAVE #NO_COPY the index of this column in the table
   bool			is_matrix;
   // #READ_ONLY #SAVE #SHOW 'true' if the cell is a matrix, not a scalar
   MatrixGeom		cell_geom;
